[#22] Move function to form util
